MAPREDUCE-714. JobConf.findContainingJar unescapes unnecessarily on Linux
Description: In JobConf.findContainingJar, the path name is decoded using URLDecoder.decode(...). This was done by Doug in r381794 (commit msg "Un-escape containing jar's path, which is URL-encoded.  This fixes things primarily on Windows, where paths are likely to contain spaces.") Unfortunately, jar paths do not appear to be URL encoded on Linux. If you try to use "hadoop jar" on a jar with a "+" in it, this function decodes it to a space and then the job cannot be submitted.
Reason: Cloudera-based packages include a '+' in the filename; Hadoop's URL escaper will not
properly handle jar filenames with a '+' without this patch.
Author: Todd Lipcon
Ref: UNKNOWN

commit d9767d2cefab288e581732f71779f3ce8e3267e4
Author: Todd Lipcon <todd@cloudera.com>
Date:   Mon Jul 6 19:36:11 2009 -0700

    MAPREDUCE-714: Fix JobConf.findContainingJars to work with jars with + in the name