HADOOP-5967. Sqoop should only use a single map task
Description: The current DBInputFormat implementation uses SELECT ... LIMIT ... OFFSET statements to
read from a database table. This actually results in several queries all accessing the same table at
the same time. Most database implementations will actually use a full table scan for each such
query, starting at row 1 and scanning down until the OFFSET is reached before emitting data to the
client. The upshot of this is that we see O(n^2) performance in the size of the table when using a
large number of mappers, when a single mapper would read through the table in O(n) time in the number of rows.

<p>This patch sets the number of map tasks to 1 in the MapReduce job sqoop launches.</p>
Reason: Performance improvement
Author: Aaron Kimball
Ref: UNKNOWN

commit 410db7130a8e85ceed46850f73e74f480d45994e
Author: Aaron Kimball <aaron@cloudera.com>
Date:   Thu Jul 23 16:10:21 2009 -0700

    HADOOP-5967: Sqoop should only use a single map task