MAPREDUCE-674. Sqoop should allow a "where" clause to avoid having to export entire tables
Description: Sqoop currently only exports at the granularity of a table.  This doesn't work well on systems with large tables, where the overhead of performing a full dump each time is significant.  Allowing the user to specify a where clause is a relatively simple task which will give Sqoop a lot more flexibility.
Reason: New feature
Author: Kevin Weil
Ref: UNKNOWN