MAPREDUCE-685. Sqoop will fail with OutOfMemory on large tables using mysql
Description: The default MySQL JDBC client behavior is to buffer the entire ResultSet in the client before allowing the user to use the ResultSet object. On large SELECTs, this can cause OutOfMemory exceptions, even when the client intends to close the ResultSet after reading only a few rows. The MySQL ConnManager should configure its connection to use row-at-a-time delivery of results to the client.
Reason: bugfix / scalability improvement
Author: Aaron Kimball
Ref: UNKNOWN