HADOOP-6148. Implement a pure Java CRC32 calculator
Description: We've seen a reducer writing 200MB to HDFS with replication = 1 spending a long time in crc calculation. In particular, it was spending 5 seconds in crc calculation out of a total of 6 for the write. I suspect that it is the java-jni border that is causing us grief.

This outperforms java.util.zip.CRC32.
Reason: Performance improvement
Author: Scott Carey and Todd Lipcon
Ref: UNKNOWN