MAPREDUCE-825. JobClient completion poll interval of 5s causes slow tests in local mode
Description: The JobClient.NetworkedJob.waitForCompletion() method polls for job completion every 5 seconds. When running a set of short tests in pseudo-distributed mode, this is unnecessarily slow and causes lots of wasted time. When bandwidth is not scarce, setting the poll interval to 100 ms results in a 4x speedup in some tests.  This interval should be parametrized to allow users to control the interval for testing purposes.
Reason: Test performance improvement
Author: Aaron Kimball
Ref: UNKNOWN