HDFS-528. Add ability for safemode to wait for a minimum number of live datanodes
Description: When starting up a fresh cluster programatically, users often want to wait until DFS is "writable" before continuing in a script. "dfsadmin -safemode wait" doesn't quite work for this on a completely fresh cluster, since when there are 0 blocks on the system, 100% of them are accounted for before any DNs have reported.

<p>This JIRA is to add a command which waits until a certain number of DNs have reported as alive to the NN.</p>
Reason: New feature
Author: Todd Lipcon
Ref: UNKNOWN