HADOOP-6400. Log errors getting Unix UGI
Description: For various reasons, the calls out to `whoami` and `id` can fail when trying to get the unix UGI information. Currently it silently ignores failures and uses the default DrWho/Tardis ugi. This is extremely confusing for users - we should log the exception at warn level when the shell execs fail.
Reason: Debug logging improvement
Author: Todd Lipcon
Ref: UNKNOWN