MAPREDUCE-1036. An API Specification for Sqoop
Description: Over the last several months, Sqoop has evolved to a state that is functional and has room for extensions. Developing extensions requires a stable API and documentation. I am attaching to this ticket a description of Sqoop's design and internal APIs, which include some open questions. I would like to solicit input on the design regarding these open questions and standardize the API.
Reason: Documentation
Author: Aaron Kimball
Ref: UNKNOWN