HDFS-861. fuse-dfs does not support O_RDWR
Description: Some applications (for us, the big one is rsync) will open a file in read-write mode when it really only intends to read xor write (not both).  fuse-dfs should try to not fail until the application actually tries to write to a pre-existing file or read from a newly created file.
Reason: bugfix
Author: Brian Bockelman
Ref: UNKNOWN