HADOOP-6503. contrib projects should pull in the ivy-fetched libs from the root project
Description: On branch-20 currently, I get an error just running "ant contrib -Dtestcase=TestHdfsProxy". In a full "ant test" build sometimes this doesn't appear to be an issue. The problem is that the contrib projects don't automatically pull in the dependencies of the "Hadoop" ivy project. Thus, they each have to declare all of the common dependencies like commons-cli, etc. Some are missing and this causes test failures.
Reason: Build system improvement
Author: Todd Lipcon
Ref: UNKNOWN