HADOOP-4368. Superuser privileges required to do "df"
Description: super user privileges are required in DFS in order to get the file system statistics (FSNamesystem.java, getStats method).  This means that when HDFS is mounted via fuse-dfs as a non-root user, "df" is going to return 16exabytes total and 0 free instead of the correct amount.

<p>As far as I can tell, there's no need to require super user privileges to see the file system size (and historically in Unix, this is not required).</p>

<p>To fix this, simply comment out the privilege check in the getStats method.</p>
Reason: Usability improvement
Author: Craig Macdonald
Ref: UNKNOWN