MAPREDUCE-1394. Sqoop generates incorrect URIs in paths sent to Hive
Description: Hive used to require a ':8020' in HDFS URIs used with LOAD DATA statements, even though the normalized form of such a URI does not contain an explicit port number (since 8020 is the default port). Sqoop matched this by hacking the URI strings it forwarded to Hive.

<p>Hive fixed this bug a while ago &#8211; Sqoop should catch up.</p>
Reason: bugfix (compatibility)
Author: Aaron Kimball
Ref: UNKNOWN