MAPREDUCE-707. Provide a jobconf property for explicitly assigning a job to a pool
Description: A common use case of the fair scheduler is to have one pool per user, but then to define some special pools for various production jobs, import jobs, etc. Therefore, it would be nice if jobs went by default to the pool of the user who submitted them, but there was a setting to explicitly place a job in another pool. Today, this can be achieved through a sort of trick in the JobConf:

<div class="code panel" style="border-width: 1px;"><div class="codeContent panelContent">
<pre class="code-java">&lt;property&gt;
  &lt;name&gt;mapred.fairscheduler.poolnameproperty&lt;/name&gt;
  &lt;value&gt;pool.name&lt;/value&gt;
&lt;/property&gt;

&lt;property&gt;
  &lt;name&gt;pool.name&lt;/name&gt;
  &lt;value&gt;${user.name}&lt;/value&gt;
&lt;/property&gt;</pre>
</div></div>

<p>This JIRA proposes to add a property called mapred.fairscheduler.pool that allows a job to be placed directly into a pool, avoiding the need for this trick.</p>
Reason: Configuration improvement
Author: Alan Heirich
Ref: UNKNOWN