MAPREDUCE-1436. Deadlock in preemption code in fair scheduler
Description: In testing the fair scheduler with preemption, I found a deadlock between updatePreemptionVariables and some code in the JobTracker. This was found while testing a backport of the fair scheduler to Hadoop 0.20, but it looks like it could also happen in trunk and 0.21. Details are in a comment below.
<hr/>
The fair scheduler introduces a potential jobtracker deadlock which
was fixed on trunk by MAPREDUCE-870. This patch adjusts the locking
in 0.20-based MapReduce to prevent this condition.

Reason: bugfix (deadlock)
Author: Matei Zaharia
Ref: UNKNOWN