HADOOP-6522. Fix decoding of codepoint zero in UTF8.
Description: TestUTF8 is actually flaky. It generates 10 random strings to run the test on. If you change this number to 100000 it fails every time. The problem is that the null character (codepoint zero) was correctly encoded but incorrectly decoded. I've attached a patch that fixes this and increases the size of the tests so that problems like this will likely be discovered sooner.

Reason: Bugfix to UTF8
Author: Eli Collins
Ref: CDH-718