HADOOP-6667. RPC.waitForProxy should retry through NoRouteToHostException.
Description: RPC.waitForProxy already loops through ConnectExceptions, but NoRouteToHostException is not a subclass of ConnectException. In the case that the NN is on a VIP, the No Route To Host error is reasonably common during a failover, so we should retry through it just the same as the other connection errors.

Reason: Improvement
Author: Eli Collins
Ref: CDH-907