HADOOP-4885. Try to restore failed replicas of Name Node storage (at checkpoint time).
Description: If one of the replicas of the NameNode storage fails for whatever reason (for example temporarily failure of NFS) this Storage object is removed from the list of storage objects forever. It can be added back only on restart of the NameNode. We propose to check the status of a failed storage on every checkpoint and if it becomes valid - try to restore the edits and fsimage.

Reason: Improvement
Author: Eli Collins
Ref: CDH-473