HADOOP-6723. Unchecked exceptions thrown in IPC connection should not orphan clients.
Description: If the server sends back some malformed data, for
example, receiveResponse() can end up with an incorrect call ID. Then,
when it tries to find it in the calls map, it will end up with null
and throw NPE in receiveResponse. This isn't caught anywhere, so the
original IPC client ends up hanging forever instead of catching an
exception. Another example is if the writable implementation itself
throws an unchecked exception or OOME.

We should catch Throwable in Connection.run() and shut down the
connection if we catch one.

Reason: Bug fix
Author: Eli Collins
Ref: CDH-1131