MAPREDUCE-1489. DataDrivenDBInputFormat should not query the database when generating only one split
Description: DataDrivenDBInputFormat runs a query to establish bounding values for each split it
generates; but if it's going to generate only one split (mapreduce.job.maps == 1), then there's no
reason to do this. This will remove overhead associated with a single-threaded import of a
non-indexed table since it avoids a full table scan.
Reason: Improvement
Author: Aaron Kimball
Ref: CDH-1431