MAPREDUCE-1213. TaskTrackers restart is very slow because it deletes distributed cache directory synchronously
Description: We are seeing that when we restart a tasktracker, it tries to recursively delete all
the file in the distributed cache. It invoked FileUtil.fullyDelete() which is very very slow. This
means that the TaskTracker cannot join the cluster for an extended period of time (upto 2 hours for
us). The problem is acute if the number of files in a distributed cache is a few-thousands.
Reason: Performance
Author: Zheng Zhao
Ref: CDH-495