MAPREDUCE-739. Allow relative paths to be created inside archives.
Description: Allow creating archives with relative paths with a -p option on
the command line.  Archives currently stores the full path from the input
sources – since it allows multiple sources and regular expressions as inputs.
So the created archives have the full path of the input sources.  This is un
intuitive and a user hassle. We should get rid of it and allow users to say
that the created archive should be relative to some absolute path and throw an
exception if the input does not confirm to the relative absolute path.
Reason: Improvement
Author: Mahadev konar
Ref: CDH-501