HDFS-1377. Quota bug for partial blocks allows quotas to be violated.
There may be a delta in FSDirectory#replaceNode even with identical
blocks because INode#diskspaceConsumed rounds up the size of the last
block if newnode is under construction. This causes us to incorrectly
reduce the space consumed for quota accounting. Looking at uses of
this functions oldnode and newnode should always have the same blocks,
therefore we should not expect a delta here.

Reason: Bug
Author: Eli Collins
Ref: CDH-2092