Amend HADOOP-7093. Revert incompatible change in semantics of HttpServer
The original backport pulled in part of HADOOP-6461, which changed the way the
"webapps" directory is located on the classpath. This broke HBase's ability
to locate its UIs. In order to avoid having to patch HBase in CDH, this
patch reverts that part of the change and works around the issue in the tests
a different way.

Reason: Should work with upstream HBase
Author: Todd Lipcon
Ref: CDH-2635