Amend MAPREDUCE-2234. Previous patch resulted in too many ls -l calls during heartbeats
The previous commit under this JIRA changed the checkLocalDirs function to use
the checkDir() function that takes a permission. This is fine at startup, but
is expensive since it results in an `ls -l` fork for every local directory.
This happens on every heartbeat and is not necessary. This patch amends
the function to only use this form of checkDir() at start time, and otherwise
just use the less expensive native java calls.

Author: Todd Lipcon
Ref: CDH-2780