MAPREDUCE-2472. Extra whitespace in mapred.child.java.opts breaks JVM initialization.
When creating taskjvm.sh, we split mapred.child.java.opts on " " and
then create a quoted argument for each of those results. So, if you
have an extra space anywhere in this configuration, you get an
argument '' in the child command line, which the JVM interprets as an
empty class name. This results in a ClassNotFoundException and the
task cannot run.

Reason: Bug
Author: Aaron T. Myers
Ref: CDH-3152