HDFS-2186. DN volume failures on startup are not counted.
Volume failures detected on startup are not currently counted/reported
as such. Eg if you have configured 4 volumes, 2 tolerated failures,
and you start a DN with two failed volumes it will come up and report
(to the NN) no failed volumes. The DN will still be able to tolerate 2
additional volume failures (ie it's OK with no valid volumes
remaining). The intent of the volume failure toleration config value
is that if more than this # of volumes of the total set of configured
volumes have failed the DN should shutdown, therefore volume failures
detected on startup should count against this quota.

Reason: Bug
Author: Eli Collins
Ref: CDH-3371