HDFS-1779. Fix a bug regarding recovery of blocks being written while NN restarts
This patch adds a new RPC 'blocksBeingWrittenReport()' which the DN calls
on startup and when it re-connects to a restarted NameNode. This reports
all blocks currently under construction, so the NN can re-add them to
the targets list for a block if necessary.

Reason: avoid HBase data-loss scenario when NN crashes
Author: Hairong Kuang
Ref: CDH-3507