MAPREDUCE-3289. Use fadvise in the TaskTracker's MapOutputServlet.
The TaskTracker now uses the posix_fadvise syscall to page in map output
before serving it to the reducers. After serving the output, it evicts
it from the buffer cache since it will not be read again in the majority
of cases.

This new behavior can be disabled by setting mapred.tasktracker.shuffle.fadvise
to false.

This patch differs from the upstream version since the upstream version applies
to the NodeManager in MR2.

Reason: Low-risk performance improvement
Author: Todd Lipcon
Ref: CDH-3818