MAPREDUCE-2777. Adds cumulative cpu usage and total heap usage to task counters. Backport MAPREDUCE-220 and MAPREDUCE-2469.
MAPREDUCE-220. Collecting cpu and memory usage for MapReduce tasks.

It would be nice for TaskTracker to collect cpu and memory usage for
individual Map or Reduce tasks over time.

MAPREDUCE-2469. Task counters should also report the total heap usage
of the task.

Currently, the task counters report VSS and RSS usage of the task. The
task counter should also report the total heap usage of the task
also. The task might be configured with a max heap size of M but the
task's total heap usage might only be H, where H < M. In such a case,
knowing only M doesn't provide a complete picture of the task's memory
usage.

Author: Scott Chen, Amar Kamat
Ref: CDH-1458