HDFS-2701. Cleanup FS* processIOError methods.
Let's rename the various "processIOError" methods to be more
descriptive. The current code makes it difficult to identify and
reason about bug fixes. While we're at it let's remove "Fatal" from
the "Unable to sync the edit log" log since it's not actually a fatal
error (this is confusing to users). And 2NN "Checkpoint done" should
be info, not a warning (also confusing to users).

Reason: Improvement
Author: Eli Collins
Ref: CDH-3921