MR1: CLOUDERA-BUILD. Preliminary version of MAPREDUCE-5707. ApplicationClientProtocolPBClientImpl (and JobClient) does not allow to set rpcTimeout
This commit adds the ability to enable rpc timeout for the job client.
The approach taken was to add a new configuration parameter
"mapreduce.jobclient.rpc.timeout". By default it is set to -1 (no
timeout) to maintain backword compatibility. If set to some positive
integer it will be used as the rpcTimeout in the underlying ipc Client.

Testing done: New unit-tests. TestRpcTimeout hanged forever before
this commit and now passes.