CDH-59917: make getJNIEnv visible in libhdfs
In hadoop versions >= 2.7, getJNIEnv is no longer visible
in libhdfs. Impala relies on getJNIEnv to interact with the
in-process JVM. This is currently blocking our frontend
tests as well as our ability to link Impala with shared
libraries.

IMPALA-2029 tracks Impala's work to eliminate our dependency
on getJNIEnv. However, this is currently blocked by
HDFS-12628, where libhdfs can crash when a client maintains
its own JNIEnv.

This is a workaround to get C6 tests running. It can be
removed when HDFS-12628 is resolved and Impala has
fixed IMPALA-2029.

Change-Id: Ie38fe2e894c1ad229f1fbaf237d9fd82a5c48253