CLOUDERA-BUILD CDH-65543 Use branch.sh to update lib.sh
The branching script performs a raw string replacement on each
component based on the starting/target branches. So for example,
branching from 6.x-SNAPSHOT to 6.0.x-SNAPSHOT results in modifying
all files that match 6\.x-SNAPSHOT to 6.0.x-SNAPSHOT.

This fails to catch some scripts in hadoop. For example
cloudera/lib.sh contains references to 6.x (which the regex above
does not catch)

If an executable script cloudera/branch.sh exists, the script will be
executed instead of the string replacement.

More info here:
http://github.mtv.cloudera.com/CDH/cdh/blob/cdh6.x/README_cauldron.md#branching-and-branch-names

Change-Id: Id8417dd4edcd7095ce7ba7639a95b6e371fab74e