CLOUDERA-BUILD CDH-67049 Using jackson versions from the root pom
==C5_APPROVED_BUGFIX==

In this commit we use the versions of jackson that are supplied in the
root pom. Previously they were hard coded. This change needs to go in
before we bump up the version in the root pom, otherwise there are
packaging failures in hadoop.

This has been tested thanks to the efforts of Csaba Fulop, who created
a special set of jenkins jobs at

http://golden.jenkins.cloudera.com/job/CDH-MYODER5.16.0-Packaging-Hadoop/

specifically for this purpose. We have two green runs there.

(cherry picked from commit 514aeca4f89dd116266880d9b1682f162f869696)

Change-Id: Ic53c17fc599c2aa34f012d34b5bb17ff94579414