CDH-61933: Prevent excessive logging of UnsupportedOperationException
This change prevents excessive logging of UnsupportedOperationException in environment
where buffered read is not supported like S3. The problem happens in hdfsOpenImp() where
a test read is done to see if the direct read is supported or not.In an
environment/storage where buffered read is not supported, every open will log an
exception UnsupportedOperationException, which results in flooding of error logs.

Change-Id: I0ad0fd0a2680793fe74821b97568a3038ea08dfd